Fix (finally) for token bucket class.
The token bucket has always had issues. I eventually managed to figure
out the underlying cause. Updates were originally done by only updating
when we were out of tokens and waiting for more. This implementation had
a fatal flaw when it came to bursts of requests.

For example: if a bucket was created with 1 token at a rate of 1 per
second:
```
t = TokenBucket(1, 1)
time.sleep(2)
start = time.now()
await t.get_token()
await t.get_token()
print(time.time() - start)

>>> 0.000629425048828125
```

First the original token is consumed. Next a second request comes in for
the now empty bucket. The bucket sees that the last update was 2 seconds
in the past, so it refills the bucket, which is immediatly comsumed.
Because of the delay, the original implementation allows requests to
burst up to twice the intended rate.

The fix for this is to update the bucket timestamp on access, not just
when it's empty.